+++ CHANGELOG.md
@@ -7,5 +7,5 @@
 - Basic library function for parsing changelogs using `changelog-parser`
 - Determine next version numbers
 - Typed change types based on https://keepachangelog.com
-- Basic CLI to print the changelog to `stdout` in "plain text"
+- Basic CLI to print the changelog to `stdout` in "plain text" (all or latest release) and Android string resource xml
 - Formatter function and options to describe how to format a changelog to text